Transfers: filter unsupported transfertools early
Instead of doing it near the very end, after path computation,
perform it at the start of request handling.
